Rotate by arrow keys.

+++ src/Point.cpp
@@ -1,6 +1,6 @@
 #include "Point.h"
 
-Point::Point( int x, int y ) {
+Point::Point( double x, double y ) {
   setX(x);
   setY(y);
 }
@@ -9,10 +9,10 @@ Point::Point() : Point( 0, 0 ) {}
 
 Point::~Point() {}
 
-void Point::setX( int x ) { this->x = x; }
+void Point::setX( double x ) { this->x = x; }
 
-void Point::setY( int y ) { this->y = y; }
+void Point::setY( double y ) { this->y = y; }
 
-int Point::getX() { return x; }
+double Point::getX() { return x; }
 
-int Point::getY() { return y; }
+double Point::getY() { return y; }
